Optimize the reading page
