+++ app/build.gradle
@@ -166,9 +166,9 @@ dependencies {
     // https://developer.android.com/jetpack/androidx/releases/compose-material
     implementation "androidx.compose.material:material:$compose"
     implementation "androidx.compose.material:material-icons-extended:$compose"
+    debugImplementation "androidx.compose.ui:ui-tooling:$compose"
     implementation "androidx.compose.ui:ui-tooling-preview:$compose"
     androidTestImplementation "androidx.compose.ui:ui-test-junit4:$compose"
-    debugImplementation "androidx.compose.ui:ui-tooling:$compose"
 
     // hilt
     implementation "androidx.hilt:hilt-work:1.0.0"
