+++ app/src/main/java/me/ash/reader/ui/ext/LazyListStateExt.kt
@@ -1,8 +1,7 @@
 package me.ash.reader.ui.ext
 
 import androidx.compose.foundation.lazy.LazyListState
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
+import androidx.compose.runtime.*
 import androidx.paging.compose.LazyPagingItems
 import kotlin.math.abs
 
@@ -27,4 +26,34 @@ fun <T : Any> LazyPagingItems<T>.rememberLazyListState(): LazyListState {
         // Return rememberLazyListState (normal case).
         else -> androidx.compose.foundation.lazy.rememberLazyListState()
     }
+}
+
+/**
+ * TODO: To be improved
+ *
+ * Returns whether the LazyListState is currently in the
+ * downward scrolling state.
+ */
+@Composable
+fun LazyListState.isScrollDown(): Boolean {
+    var isScrollDown by remember { mutableStateOf(false) }
+    var preItemIndex by remember { mutableStateOf(0) }
+    var preScrollStartOffset by remember { mutableStateOf(0) }
+
+    LaunchedEffect(this) {
+        snapshotFlow { isScrollInProgress }.collect {
+            if (isScrollInProgress) {
+                isScrollDown = when {
+                    firstVisibleItemIndex > preItemIndex -> true
+                    firstVisibleItemScrollOffset < preItemIndex -> false
+                    else -> firstVisibleItemScrollOffset > preScrollStartOffset
+                }
+            } else {
+                preItemIndex = firstVisibleItemIndex
+                preScrollStartOffset = firstVisibleItemScrollOffset
+            }
+        }
+    }
+
+    return isScrollDown
 }
