+++ app/src/main/java/me/ash/reader/ui/page/home/reading/Header.kt
@@ -1,62 +1,67 @@
 package me.ash.reader.ui.page.home.reading
 
-import android.content.Intent
-import android.net.Uri
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
-import me.ash.reader.data.entity.ArticleWithFeed
 import me.ash.reader.ui.ext.formatAsString
+import me.ash.reader.ui.ext.openURL
 import me.ash.reader.ui.ext.roundClick
+import java.util.*
 
 @Composable
 fun Header(
-    articleWithFeed: ArticleWithFeed,
+    feedName: String,
+    title: String,
+    author: String? = null,
+    link: String? = null,
+    publishedDate: Date,
 ) {
     val context = LocalContext.current
+    val dateString = remember(publishedDate) {
+        publishedDate.formatAsString(context, atHourMinute = true)
+    }
 
     Column(
         modifier = Modifier
             .fillMaxWidth()
             .roundClick {
-                articleWithFeed.article.link.let {
-                    if (it.isNotEmpty()) {
-                        context.startActivity(
-                            Intent(Intent.ACTION_VIEW, Uri.parse(articleWithFeed.article.link))
-                        )
-                    }
-                }
+                context.openURL(link)
             }
             .padding(12.dp)
     ) {
         Text(
-            text = articleWithFeed.article.date.formatAsString(context, atHourMinute = true),
-            color = MaterialTheme.colorScheme.outline.copy(alpha = 0.7f),
+            modifier = Modifier.alpha(0.7f),
+            text = dateString,
+            color = MaterialTheme.colorScheme.outline,
             style = MaterialTheme.typography.labelMedium,
         )
         Spacer(modifier = Modifier.height(4.dp))
         Text(
-            text = articleWithFeed.article.title,
+            text = title,
             color = MaterialTheme.colorScheme.onSurface,
             style = MaterialTheme.typography.headlineLarge,
         )
         Spacer(modifier = Modifier.height(4.dp))
-        articleWithFeed.article.author?.let {
+        author?.let {
             if (it.isNotEmpty()) {
                 Text(
+                    modifier = Modifier.alpha(0.7f),
                     text = it,
-                    color = MaterialTheme.colorScheme.outline.copy(alpha = 0.7f),
+                    color = MaterialTheme.colorScheme.outline,
                     style = MaterialTheme.typography.labelMedium,
                 )
             }
         }
         Text(
-            text = articleWithFeed.feed.name,
-            color = MaterialTheme.colorScheme.outline.copy(alpha = 0.7f),
+            modifier = Modifier.alpha(0.7f),
+            text = feedName,
+            color = MaterialTheme.colorScheme.outline,
             style = MaterialTheme.typography.labelMedium,
         )
     }
